For module-only consumers, avoid depending on nixpkgs
This seems odd, but it works since input dependencies are lazily fetched
that means a downstream repo, that only consumes the modules output
(like for example `divnix/digga`), does never fetch nor instantiate
nixpkgs proper (which would be costly).

This enables `divnix/digga` to continue to _not_ depend on `nixpkgs`
for operation.

(it still depends on `nixpkgs` as a strict development dependency, but
beacause of the black magic which I describe above, no end user
will ever fetch it's depenency on `nixpkgs`.

So, here, we want to implement a similar trick to avoid acquireing deps
where they should net be acquired.