lockfile: Update inputs
This commit updates the flake inputs of nixos-generators, as they are
currently very outdated, which can cause a range of issues for users
of nixos-generators who does not override these inputs to newer
versions themselves.

The change was tested by running `nix flake show` to verify that the
nixosModules outputs are generated correctly with the new version of
`nixpkgs.lib`.