Fix the extension breaking the find query.
In af3d3a7538d097ec927238d868a04e765ee07ed6, filename was renamed to fileExtension and used in find -name. This breaks kexec as it used to have the actual file name glob in there.

Given it's the only file produced it seems safe to just remove the path component.