iso: default to zstd compression
For most users of nixos-generate zstd will be the better choice as
it will build and boot images much faster. Often users have to rebuild
images a couple of times before they are happy with the result.
NixOS might be better of with xz as it saves download and storage cost
for images.