Fix nixosModules example
Previously, it did not compile.  The goal of the section appears to have been
to describe how to make a nixosModule that will imbue format magic upon
configurations referencing it.  With that in mind, I also changed comments and
attribute names to make it more clear that the nixosConfiguration is using the
nixosModule rather than integral to it or vice versa.